requirements.txt
numpy
matplotlib
pyyaml
zarr
torchvision
torchinfo
tqdm
torchmetrics
pandas
scikit-learn
scipy
seaborn
timm